Update access to pages Only authorized access is possible now

+++ mantis/views/corpus.py
@@ -5,6 +5,7 @@ from sqlalchemy.exc import SQLAlchemyError, IntegrityError
 
 from mantis.model import Corpus, Message, Document, serialize
 from mantis.utils import ctx
+from mantis.views.login import check_login
 
 __all__ = [
     'page',
@@ -27,63 +28,50 @@ def serialize_corpora(e):
 @page.route('/', methods=['GET'])
 def render_corpus():
     with ctx.SessionContext() as app:
-        corpora = serialize_corpora(app)
-        html = render_template('pages/corpus.html', messages=app.messages, corpora=corpora, mode='view')
-        app.clear('messages')
-        return html
+        user = check_login(app)
+        if user:
+            corpora = serialize_corpora(app)
+            html = render_template('pages/corpus.html', messages=app.messages, corpora=corpora, mode='view')
+            app.clear('messages')
+        else:
+            html = redirect(url_for('index_page.render_index'))
+    return html
 
 
 @page.route('/create', methods=['POST'])
 def render_corpus_add():
     with ctx.SessionContext() as app:
-        corpus = Corpus(name=request.form['corpus_name'])
-        try:
-            app.database.add(corpus)
-            app.database.commit()
-            app.messages.append(Message(Message.Type.SUCCESS, 'Corpus created successfully'))
-        except IntegrityError as err:
-            app.database.rollback()
-            if "Duplicate entry '{}' for key 'name'".format(corpus.name) in str(err):
-                msg = 'Corpus name already exists \'{}\''.format(corpus.name)
-                app.messages.append(Message(Message.Type.ERROR, msg))
-            else:
+        user = check_login(app)
+        if user:
+            corpus = Corpus(name=request.form['corpus_name'])
+            try:
+                app.database.add(corpus)
+                app.database.commit()
+                app.messages.append(Message(Message.Type.SUCCESS, 'Corpus created successfully'))
+            except IntegrityError as err:
+                app.database.rollback()
+                if "Duplicate entry '{}' for key 'name'".format(corpus.name) in str(err):
+                    msg = 'Corpus name already exists \'{}\''.format(corpus.name)
+                    app.messages.append(Message(Message.Type.ERROR, msg))
+                else:
+                    msg = 'Please contact admin for further assistance'
+                    app.messages.append(Message(Message.Type.ERROR, msg))
+            except SQLAlchemyError as _:
                 msg = 'Please contact admin for further assistance'
                 app.messages.append(Message(Message.Type.ERROR, msg))
-        except SQLAlchemyError as _:
-            msg = 'Please contact admin for further assistance'
-            app.messages.append(Message(Message.Type.ERROR, msg))
-        return redirect(url_for('.render_corpus'))
+    return redirect(url_for('.render_corpus'))
 
 
 @page.route('/delete', methods=['GET'])
 def render_corpus_delete():
     id_ = request.args.get('id', None, int)
     with ctx.SessionContext() as app:
-        if id_ is not None:
-            corpus = app.database.query(Corpus).get(id_)
-            if corpus is not None:
-                app.database.delete(corpus)
-                app.database.commit()
-                msg = 'Corpus with ID \'{}\' deleted successfully.'.format(id_)
-                app.messages.append(Message(Message.Type.SUCCESS, msg))
-            else:
-                msg = 'No such corpus with ID \'{}\' to delete.'.format(id_)
-                app.messages.append(Message(Message.Type.ERROR, msg))
-        else:
-            msg = 'Invalid value for parameter ID. Found \'{}\' expected an integer.'.format(id_)
-            app.messages.append(Message(Message.Type.ERROR, msg))
-    return redirect(url_for('.render_corpus'))
-
-
-@page.route('/edit', methods=['GET', 'POST'])
-def render_corpus_edit():
-    with ctx.SessionContext() as app:
-        id_ = request.args.get('id', None, int)
-        if id_ is not None:
-            if request.method == 'POST':
+        user = check_login(app)
+        if user:
+            if id_ is not None:
                 corpus = app.database.query(Corpus).get(id_)
                 if corpus is not None:
-                    corpus.name = request.form['corpus_name']
+                    app.database.delete(corpus)
                     app.database.commit()
                     msg = 'Corpus with ID \'{}\' deleted successfully.'.format(id_)
                     app.messages.append(Message(Message.Type.SUCCESS, msg))
@@ -91,16 +79,39 @@ def render_corpus_edit():
                     msg = 'No such corpus with ID \'{}\' to delete.'.format(id_)
                     app.messages.append(Message(Message.Type.ERROR, msg))
             else:
-                corpora = serialize_corpora(app)
-                html = render_template(
-                    'pages/corpus.html', messages=app.messages, corpora=corpora, mode='edit', edit_id=id_
-                )
-                app.clear('messages')
-                return html
-        else:
-            msg = 'Invalid value for parameter ID. Found \'{}\' expected an integer.'.format(id_)
-            app.messages.append(Message(Message.Type.ERROR, msg))
-        return redirect(url_for('.render_corpus'))
+                msg = 'Invalid value for parameter ID. Found \'{}\' expected an integer.'.format(id_)
+                app.messages.append(Message(Message.Type.ERROR, msg))
+    return redirect(url_for('.render_corpus'))
+
+
+@page.route('/edit', methods=['GET', 'POST'])
+def render_corpus_edit():
+    with ctx.SessionContext() as app:
+        user = check_login(app)
+        if user:
+            id_ = request.args.get('id', None, int)
+            if id_ is not None:
+                if request.method == 'POST':
+                    corpus = app.database.query(Corpus).get(id_)
+                    if corpus is not None:
+                        corpus.name = request.form['corpus_name']
+                        app.database.commit()
+                        msg = 'Corpus with ID \'{}\' deleted successfully.'.format(id_)
+                        app.messages.append(Message(Message.Type.SUCCESS, msg))
+                    else:
+                        msg = 'No such corpus with ID \'{}\' to delete.'.format(id_)
+                        app.messages.append(Message(Message.Type.ERROR, msg))
+                else:
+                    corpora = serialize_corpora(app)
+                    html = render_template(
+                        'pages/corpus.html', messages=app.messages, corpora=corpora, mode='edit', edit_id=id_
+                    )
+                    app.clear('messages')
+                    return html
+            else:
+                msg = 'Invalid value for parameter ID. Found \'{}\' expected an integer.'.format(id_)
+                app.messages.append(Message(Message.Type.ERROR, msg))
+    return redirect(url_for('.render_corpus'))
 
 
 def allowed_file(filename):
@@ -110,31 +121,33 @@ def allowed_file(filename):
 @page.route('/add', methods=['POST'])
 def render_corpus_document_add():
     with ctx.SessionContext() as app:
-        id_ = request.args.get('id', None, int)
-        if id_ is not None:
-            if 'file' in request.files:
-                file = request.files['file']
-                if file.filename != '':
-                    if file and allowed_file(file.filename):
-                        file_data = file.read().decode('utf-8')  # latin-1
-                        for line in file_data.split('\n'):
-                            text = line.strip()
-                            document = Document(corpus_id=id_, text=text)
-                            app.database.add(document)
-                        app.database.commit()
-                        msg = 'Documents added to Corpus with ID = `{}` successfully'.format(id_)
-                        app.messages.append(Message(Message.Type.SUCCESS, msg))
+        user = check_login(app)
+        if user:
+            id_ = request.args.get('id', None, int)
+            if id_ is not None:
+                if 'file' in request.files:
+                    file = request.files['file']
+                    if file.filename != '':
+                        if file and allowed_file(file.filename):
+                            file_data = file.read().decode('utf-8')  # latin-1
+                            for line in file_data.split('\n'):
+                                text = line.strip()
+                                document = Document(corpus_id=id_, text=text)
+                                app.database.add(document)
+                            app.database.commit()
+                            msg = 'Documents added to Corpus with ID = `{}` successfully'.format(id_)
+                            app.messages.append(Message(Message.Type.SUCCESS, msg))
+                        else:
+                            msg = 'File not allowed in the system'
+                            app.messages.append(Message(Message.Type.ERROR, msg))
                     else:
-                        msg = 'File not allowed in the system'
+                        msg = 'No selected file'
                         app.messages.append(Message(Message.Type.ERROR, msg))
                 else:
-                    msg = 'No selected file'
+                    msg = 'No file part'
                     app.messages.append(Message(Message.Type.ERROR, msg))
             else:
-                msg = 'No file part'
+                msg = 'Invalid value for parameter ID. ' \
+                      'Found \'{}\' expected an integer'.format(request.args.get('id', None))
                 app.messages.append(Message(Message.Type.ERROR, msg))
-        else:
-            msg = 'Invalid value for parameter ID. ' \
-                  'Found \'{}\' expected an integer'.format(request.args.get('id', None))
-            app.messages.append(Message(Message.Type.ERROR, msg))
     return redirect(url_for('.render_corpus'))
